(#17031) Prefer the domain component from the resolved SID
Previously, the `members_to_s` message included the SID string which was
difficult to read when the group has more than a few members.

Also, given a manifest with a fully qualified member:

    members => ['.\Administrator']`

The resulting `members changed` message contained:

    BUILTIN\.\Administrator

This is because `Win32::Security::SID#account` returns the value originally
passed to the constructor, .\Administrator. This doesn't actually affect the
group membership, since we always use the SID string in the ADSI connection
string, e.g. WinNT://S-1-5-32-*

This commit changes the message to prefer the domain component from the
resolved sid, e.g SID#domain, instead of the value the user entered, as the
former is the canonical version.

It also only prints the SID for each member when debugging is enabled.