(#19514) Fix up tests that were not providing a valid Node
Several of the tests were not providing a valid Puppet::Node object when
they called the compiler. They were either providing nil, or a mock.
This resulted in invalid calls and failing tests when the internals of
the compiler were changed. Instead of changing the internals of the
compiler for this invalid way of being called, this changes the tests
and makes the compiler assume that it is working with good parameters.