(packaging) (22726) Add saucy to default cows
Saucy comes out soon, and we would like to provide packages for that
distro. In order to do this, we need to add the saucy cow to our default
cow list, so that each new release we have, we will build and ship saucy
packages, in addition to the packages we already ship.