(#22910) Deprecate --host on puppet resource
With the networked resource access deprecated, the --host option also
needs to be deprecated so that users can easily see that it is
deprecated. Instead of using Puppet.deprecation_warning, this uses
simply Puppet.warning so that the user isn't presented with a confusing
code location on the command line. This should be ok since the
throttling of the message shouldn't be needed from the CLI environment.