(#20516) Add support for epp templates in new lexer
This adds support for epp templates to the new Lexer2 without
affecting the overall performance of the lexer.

This does not add the full epp template feature, only the support in the
lexer and the EppParser that shows how the lexer is used to perform epp
lexing when parsing.

This commit also refacores method names for error handling since
the most common way to report a lexer error is to add location
information.