(perf) Make future parser use new faster lexer
This makes the future parser use the new faster lexer.
As a consequence features provided by the old lexer has now moved
to the parser (keeping track of namestack).

Locator information is still not optimized wrt the parser as the
tokens/locator information are adapter to the "old" way
(room for performance improvements).

Tests are updated as the new lexer is both more restrictive, and in
some case more relaxed (where regular expressions may occur).