(#22918) Lock the filebucket for updates
There were 2 races in the filebucket saving logic:

  1. 2 writers could try to create the `contents` file at the same time.
     This resulted in the writer that loses the race raising an error
     because it did not have write permission to the `contents` file that
     the first writer created.
  2. 2 writers could try updating the `paths` file at the same time.
     This would result in both entering the path in the file thereby
     causing duplicate information in the file. It might also have been
     possible for the two writes to be interleaved in a manner that garbled
     the paths (for instance if the write buffers end up flushing twice in
     an interleaved fashion).

The new method (exclusively_update_file) obtains a LOCK_EX on the
`paths` file before any changes are attempted. This makes the `paths`
file the mutex, not only for updating the paths, but also setting the
contents. This requires that the filesystem supports flock.

Note: the Puppet::Settings.readwritelock method does not work as
intended. It also attempts to get an exclusive lock on a file so that
processes can update it safely, but after testing it out, it actually
does not achieve this. In writing the new method we discovered that the
exclusive lock ability is lost if the update is done using a tempfile
and atomic replace of the file that is locked. When that happens there
is an opportunity for two processes to attain locks and modify the file.
The solution to this was to *only* allow updates-in-place.