(#22180) Add integration test for puppet doc --mode rdoc
We were lacking any comprehensive testing of the interaction between
puppet doc and rdoc.  This integration test checks either the rdoc1 or
rdoc2 output (depending on which version of rdoc is present on the
system) for all of the basic module constructs we are documenting:
modules, classes, defined types, nodes, and plugin facts, types and
functions.  It tests both site manifests and modules.