(Maint) Compare filebucket contents with streams
The old code required that we completely read into memory both the current
filebucket contents and the new contents. This doubled the amount of ram
needed when checking if the new contents were the same. Instead of
reading in the entire contents, we can instead work from streams and use
the FileUtils.compare_streams method.