(perf) Speed up containment calculation by 2x
This speeds up getting all contents of a model by a factor 2.
The performance is gained by not calling the general routines since
they call through many layers with enumeration, filtering and copy &
concat until they reach an answer.

The optimized version inlines as much as possible and keeps a cache of
resulting getters per class. (The getters that produce the content).