(perf) Make location calculation lazy and pass all tests
A class (and subclasses) called Locatable were added to delay the
computation of location as long as possible.

SourcePosAdapter was modified to hold a Locatable instead of
individual measure (line, pos, etc.).

Factory was modified to make use of the new Locator when creating
source positions. It also got some additional performance tweaks.

The egrammar had a bug where it cause recursive lazy location
resolution. A located result should not be located again since it
resolves to itself.

Lexer support was modified to make use of Locatabel.

Transformer spec used an old SourcePosAdapter and required changing.