(maint) Protect against endless recursion if user calls with wrong type
When user of the function find_adapter called this method with
an array, the eContainer method returned the container of each member
of the array until it got an emtpy array. At that point recursion became
endless.

The function is not designed to be called with an array of
objects, but better to produce a nil value than causing an endless
recursion for which no stack trace can be obtained.