(#7659) Fix commentstack when parsing hashes
We need to pop from commentstack in the grammar whenever we parse a
hash; an LBRACE pushes to commentstack, yet an AST::ASTHash does not
have a doc string which would get the comment popped from the stack
again. This left an extra item on the commentstack which caused all
subsequent Lexer#getcomment to return "".

Make the test for "comments before statement" contain a more complex
example to demonstrate that nested comments should not be returned.