(#23074) Add reinventory command to puppet cert
This adds a reinventory command to the puppet cert application.

Further work on this is to refactor the use of the #apply method
found in Puppet::SSL::CertificateAuthority as it delegates to
a subclass called Interface to do the work. This is all an application
concern and it is confusing with the back and forth between the classes
"the cert application", the CA and its Interface.