(#23074) Make cert inventory not rebuild inventory.txt when adding
There is a race condition that requires protection if the inventory
performs rebuild of a missing inventory.txt file when adding a cert.
Without the protection the inventory may end up with double entries or
missing entries, and with the protection the processing of cert requests
become serialized over all processors creating a performance problem.

This change simply appends to the inventory.txt - if the file does not
exists, it is atomically created.

A users that wants to recreate the inventory should do so with the
command `puppet cert reinventory` which may be executed on an
inactive system.