(#23074) Refactor cert application's use of cert authority's interface
The code was difficult to understand as the cert application called
the apply function on the CertAuthority, which in turned created an
instance of the Interface class, which in turn called back into the
ca. As this construct was only used by the cert application, the ca's
apply method was moved to the cert application.

It is now more clear that the application interacts with the ca via
an instance of its Interface class.