Maint: Reformat indirection reference to use H4s instead of DLs for termini
A long time ago, I shortsightedly decided to use the definition list Markdown
extension in some of our docs. Time and the renderer ecosystem haven't been kind
to that syntax, so we need to get away from them if we want to be able to switch
relatively freely between various MD processors.

This commit removes them from the last embedded documentation that we still use.
The old-style type reference still uses them, but since we've stopped relying on
it and moved to our own implementation, that should be fine.