(maint) Remove large spec test block whose benefits were unclear.
This changes removes a large block of unit tests, comprising
1296 individual tests built up via 8 nested 'each' loops. Why
remove so many tests?

First, in the course of refactoring for #23081 we triggered
test failures in this block only when we also triggered a test
failure elsewhere, and conversely, it didn't trigger failures
where other individual tests did register a useful failure.

Second, several people have stared at this unit test block
and not been able to figure out what it is uniquely testing.

So in short the value of this test block is not clear, so we're
removing it.  Of course, if we hit future gaps in unit test coverage,
we'll add tests back in to cover gaps.