(#23084) Package options property for package
This changeset adds the package_options property to "package" type. It's
intended to be used mainly with FreeBSD ports to control port options on
installed packages (the options normally set with make config).

The sole implementation of port_options is delegated to providers. This
changesed only introduces port_options feature and property to package
type (this initially defines an interface between package type and
provider used to handle package_options).