(#7659) parser: fix missing puppetdoc caused by hash literals
The parser fails to associate a docstring to any defined type
or class that contained a hash literal in its head or body.

This will work:

  # Documentation here.
  class foo {
  }

These will not:

  # Documentation here.
  class foo {
    $var = {}
  }

  # Documentation here.
  define bar($options = { "key" => "value" }) {
  }

  # Documentation here.
  class foo {
    bar { "baz":
      options => {
        "apply" => true,
	"try"   => false,
      }
    }
  }

Apparently, the reason is that the lexer maintains a stack of nesting levels
of braces. When encountering a hash literal, the stack is raised, but there is
no code for lowering it again at the closing brace. It's not understand why
other pairs of braces do not exhibit the same behavior.

Popping an element whenever lexing a closing brace is not an option, because
those are plentiful in any manifest. Limiting this action to such braces that
close a hash, the lexer would need to maintain yet another stack of hash levels
it's operating in at any time.

The simpler solution is to ignore opening braces of hash literals. The lexer
can identify those by context. Hashes can appear in three contexts:
1. a variable or parameter default assignment
2. as a parameter value in a resource or class declaration
3. as a value in a selector
In the first case, it is preceded by an = equals sign, otherwise by =>

Neither token can precede a code block enclosed in braces. Therefor they
distinctly identify an opening hash. If either of them precedes an opening
brace, the lexer will now ignore it wrt. the comment stack.