(doc) Add references to RFCs that deal with CSRs
Puppet::SSL::CertificateReequest implements several RFC specified
behaviors around X509 extension requests, but unless one is intimately
familiar with the relevant RFCs (and few mortals are) it's unclear if
this is an arbitrary implementation or actual standardized behavior.
This commit adds documentation that explains the implemented behavior
and references the RFCs that is being implemented.