(#22363) First wiring of new evaluator to parser factory
This is the initial work of wiring the new evaluator into the regular
puppet logic. This is achieved via a new E4ParserAdapter that is used
together with the AST::PopsBridge classes. The ParserFactory is modified
to instantiate the new E4ParserAdapter.

Since not everything is finished, there are many tests failing, but at
this point a command like puppet apply --parser future -e 'notice hi' is
parsed and evaluated using the new evaluator.