(#23096) Explicitly clear hiera terminus between specs.
Puppet::DataBinding::Hiera stores the configured terminus in a class
instance variable, which persists between spec runs.  The
databinding/hiera_spec configures the hiera terminus to point to a
fixtures datadir with an invalid yaml file for testing.  But this was
causing a test in the parser/compiler_spec to fail when it compiled a
parameterized class resource, invoking a hiera lookup for the
parameter...which would blow up when it hit the invalid yaml.

This change explicitly clears the Puppet::DataBinding::Hiera class
instance's @hiera configuration after every hiera_spec.