(#4820) cron: forbid the mixing of special and numeric schedules
The BSD special schedules such as @reboot or @midnight and regular schedules
comprising one or more time fields such as minute or hour are mutually
exclusive. Until now it was valid to specify a special schedule alongside
numeric fields, though.

Add validation logic to the crontype to ensure that a cron resource does not
try to manage these incompatible fields together. It is acceptable to
declare should-values for the respective properties, but only with the
following limitation:

If the 'special' property is declard but not :absent,
any declared numeric schedule field must be :absent.