(maint) Remove unused OptionalAttributeOperation
The expression OptionalAttributeOperation was thought as an
instruction that would treat an assignment of :undef as
antimatter vs. setting an explicit nil. This is not possible
when using the 3x Resource class, and no concrete syntax was
introduced.

This removes it from the model and the todo note to implement it
in the evaluator.