(PUP-1158) Only output load warning when file given by user has an error
There are two optimised paths through the Augeas loading code:

  * when context parameter is given
  * when lens and incl parameters are given

In the first case, all files for the lens are still loaded, including the one
the user intends to edit.  In the second, only one file and lens is loaded.

When optimising, a warning is logged for the user if it's thought that loading
of the user's file has failed.  Previously this code would output the warning
if there were any errors at all, so a file handled by the same Augeas lens
that errored would trigger the warning, even though the targeted file was fine.

Now errors for the specific context given by the user are checked in order to
issue the warning message.