(maint) stub $CHILD_STATUS for autosign_command specs
Without this commit, the value of $CHILD_STATUS in the autosign_command specs
is not stubbed, so is dependent on the return value of the last command in the
specs. This can lead to spurious test failures because this leaks
information between tests. This commit stubs the $CHILD_STATUS global
variable so that the autosign_command tests are always isolated.