(#9546) Preserve execute permissions on 'module install'
Previously (even with the patch to 'module build') execute
permissions were stripped from everything in a module.  This
was introduced to fix the security holes discussed in #14333.
However that fix (which set all files to 644) had the unneeded
side effect of discarding the executable bit.

So this change preserves permissions from the tarball, but
removes sticky/setuid/setgid bits (which are the security hole)
and removes file write permissions (which aren't needed).  This
should address the needs of both #9546 and #14333.