(PUP-1327) add the file attributes owner, group and mode to nagios types
It's desirable to make it possible to write manifests such as

nagios_host { "hostname":
  ... # nagios properties
  group => "nagios",
  mode  => 640,
}

instead of the workaround of

nagios_host { "hostname":
  target => "/etc/nagios/hostname.cfg",
  ...
}
file { "/etc/nagios/hostname.cfg": ... }

To make this possible in a simple fashion, just add the file properties
owner, group and mode to each nagios type as parameters. If either of them
is specified, the nagios resource generates an additional file resource
and passes the parameters as should-values of the respective properties.