(maint) Delete reference grammar.ra
The file pops/parser/grammar.ra was an intermediate step between
the old and the new grammar implementation. It used the old grammar
but created the new model. It can now be removed as the new grammar
is functioning as expected. (The original grammar is still available
as an up to date reference. The copy in pops/parser was not maintained).