(#22363) Improve handling of PNilType
Without this an inference of nil / undef results in an error. Now
the String 'Undef' is used.

Also, without this, a check for assignability to Nil would result
in assignability to PObjectType (which is the widest type), and this
would be wrong. (This is quite esoteric, but added for correctness).

To prevent the type reference Undef to be interpreted as a Resource
type, it has also been added to the type parser. And as a consequence
also to the type factory.